requirements.txt
boto3==1.20.24
google_api_python_client==2.65.0
google_auth_oauthlib==0.4.6
protobuf==4.21.8
requests==2.26.0
